+++ the_problem_manager/lib/nameForm.dart
@@ -1,0 +1,62 @@
+import 'package:flutter/material.dart';
+
+import 'nameData.dart';
+
+class NameForm extends StatefulWidget {
+  @override
+  State<StatefulWidget> createState() {
+    return NameFormState();
+  }
+}
+
+class NameFormState extends State {
+  @override
+  Widget build(BuildContext context) {
+    return nameForm();
+  }
+
+  GlobalKey<FormState> formKey = GlobalKey<FormState>();
+  NameData nameData = NameData();
+
+  Widget nameForm() {
+    return Form(
+      key: formKey,
+      child: Column(
+        children: [
+          yourNameField(),
+          button(),
+        ],
+      ),
+    );
+  }
+
+  Widget yourNameField() {
+    return TextFormField(
+      keyboardType: TextInputType.name,
+      validator: (String inValue) {
+        if (inValue.length == 0) {
+          return "Please enter your name";
+        }
+        return null;
+      },
+      onSaved: (String inValue) {
+        nameData.yourName = inValue;
+      },
+      decoration: InputDecoration(
+        hintText: "add your name",
+        labelText: "Name",
+      ),
+    );
+  }
+
+  Widget button() {
+    return RaisedButton(
+      child: Text("Continue"),
+      onPressed: () {
+        if (formKey.currentState.validate()) {
+          formKey.currentState.save();
+        }
+      },
+    );
+  }
+}
